refactor: TechStackSelect components/ui/selector로 이동
